refactor(storage): extract pure helpers to _helpers.py
Moves 5 functions (_d, provider_day_bounds, status_family_for,
redact_params, _nullable_int, _nullable_float) and 2 module constants
(_PROVIDER_DAY_TZ, _REDACTED_PARAM_KEYS) to storage/_helpers.py.

External callers (sources/ohlc.py, api/client.py, api/routers/health.py,
api/routers/provider_usage.py, tests/unit/storage/test_provider_usage_helpers.py)
keep their existing import paths working via explicit re-export from
repository.py.

repository.py: ~5040 -> ~4970 lines.
